Added an interface for pushback and for the killbox.
added an interface for pushBack and killbox to be used by the player and enemies.
The inherited methods are set up on the playercontroller. Launch and BounceOff were also move to the interface.
Cleaned up the moving platform script and got rid of the waypoint path script since its not needed anymore.