Added jumping VFX, a bullet impact effect for the player, and SFX to the Wind object
The player has a particle system at their feet that creates a dust cloud when they jump.
The player controller is now designed to take in a bullletImpactFX particle system, which is a moving particle system that creates a gun impact effect where ever the player shoots.
Added a sound source to the wind which is turned off and on by triggers.