Created some scripts that will be used for data persistence
created the scripts for PlayerSaveData, SaveFileHandler, and a DataManager. There is also an Interface for persistent data, which can be implemented on any script that needs to save & load.

At this point the PlayerSaveData and Interface should be technically set up, but the functionality to read and write to/from a file is not in place yet