WIP Wall Jumping Functionality
Wall jumping uses a controller collider hit and some conditions to determine if they player is attempting a wall jump. If conditions are met then the player will jump off the wall. The direction uses the Vector3.Reflect method to get the players current direction mirrored from the collider hits normal.
(This functionality kinda works, but is not finished at the point)