Added Constant Water Source Functionality and A Constant Water Source Prefab
Constant water source fills the tank similarly to the regular fill tank functionality except its designed for floats, and takes the jetpack into account.
When the player enters a water source a public bool is triggered, which tells the constant water source functionality to run each fixed update frame. When the player exits the source the bool gets turned off.