Added Some Logic to fix the player being able to jump in the air
I added a grounded timer which resets whenever the player grounded check happens in update

This timer is used to either allow or prevent a delay for when the player can jump in the air after walking off a platform.
Any grounded jump will have use this timer to check if they are allowed to occur or not.